\name{BOXWP}
\alias{BOXWP}
\title{
Boxplots by group(s).
}
\description{This function is called by Rcmdr menu items and is not
  meant to be called by the end user.}
\usage{
BOXWP()
}
%- maybe also 'usage' for other objects documented here.
\details{
  Boxplots by group for Rcmdr using lattice graphics.
}
\value{
NULL
}
\author{
P. Svante Eriksen <svante@math.aau.dk>
}
\seealso{ \code{\link{bwplot}} }
