\name{HIST}
\alias{HIST}
\title{
Histograms by group(s).
}
\description{This function is called by Rcmdr menu items and is not
  meant to be called by the end user.}
\usage{
HIST()
}
%- maybe also 'usage' for other objects documented here.
\details{
  Histograms by group for Rcmdr using lattice graphics.
}
\value{
NULL
}
\author{
P. Svante Eriksen <svante@math.aau.dk>
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\seealso{ \code{\link{histogram}} }
