\name{QQN}
\alias{QQN}
\title{
Normal q-q-plots by group(s).
}
\description{This function is called by Rcmdr menu items and is not
  meant to be called by the end user.}
\usage{
QQN()
}
%- maybe also 'usage' for other objects documented here.
\details{
  Normal q-q-plots by group for Rcmdr using lattice graphics.
}
\value{
NULL
}
\author{
P. Svante Eriksen <svante@math.aau.dk>
}
\seealso{ \code{\link{qqmath}} }
