\name{RcmdrPlugin.plotByGroup-package}
\alias{RcmdrPlugin.plotByGroup}
\docType{package}
\title{
Rcmdr lattice Plug-In
}
\description{
This package provides an R-commander plug-in for doing some lattice graphics.
}
\details{
\tabular{ll}{
Package: \tab RcmdrPlugin.plotByGroup\cr
Type: \tab Package\cr
Version: \tab 0.1-0\cr
Date: \tab 2013-01-28\cr
License: \tab GPL 2\cr
}

}
\author{
P. Svante Eriksen <svante@math.aau.dk>

Maintainer: Ege Rubak <rubak@math.aau.dk>
}
\references{
Fox, J. (2007) Extending the R Commander by ``plug in'' packages. \emph{R News}, 7(3): 46-52.
}

\keyword{ package }
\seealso{
  \code{\link{Rcmdr}}
  \code{\link{histogram}}
  \code{\link{bwplot}}
  \code{\link{qqmath}}
}
